pyproject.toml
[tool.poetry]
name = "python-final-project"
version = "0.1.0"
description = ""
authors = ["Will Lane <williamlane923@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fire = "^0.4.0"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.5b2", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = 'python_final_project.cli:main'
